Feat/homepage : 컴포넌트 변경, 스타일 수정, 이미지 환경설정 등
