+++ frontend/src/app/hosting/privacy-type/page.tsx
@@ -29,7 +29,7 @@ export default function Page() {
 
   return (
     <main>
-      <section className="h-screen flex flex-col border border-dashed">
+      <section className="h-screen flex flex-col border/ border-dashed/">
         <div className="max-w-2xl mx-auto flex-1 py-10 flex flex-col justify-center gap-10">
           <header className="text-center">
             <h1 className="text-3xl font-bold tracking-tight">
@@ -47,7 +47,7 @@ export default function Page() {
           />
         </div>
 
-        <div className="py-10 flex justify-between border border-dashed">
+        <div className="py-10 flex justify-between border/ border-dashed/">
           <Button variant="ghost" onClick={() => router.back()}>
             이전
           </Button>
