+++ frontend/src/app/page.tsx
@@ -1,10 +1,10 @@
-import { RoomCard } from '@/components/room/RoomCard';
 import type { Room } from '@/types/room';
+import RoomCard from '@/components/room/RoomCard';
 
 async function getRooms() {
   try {
     const url = `${process.env.API_URL}/rooms`;
-    console.log({ url });
+    // console.log({ url });
     const res = await fetch(url);
     if (!res.ok) throw new Error('Failed to fetch rooms');
     return res.json();
@@ -20,17 +20,19 @@ export default async function Page() {
 
   return (
     <main>
-      <section
-        className="grid 
-        grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5
-        gap-x-5 gap-y-10 
-        sm:px-4 lg:px-4
-        pt-10
-        "
-      >
-        {rooms.map((room) => (
-          <RoomCard key={room.id} room={room} />
-        ))}
+      <section>
+        <ul
+          className="grid
+          grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5
+          gap-x-5 gap-y-10
+          sm:px-4 lg:px-4
+          pt-10
+          "
+        >
+          {rooms.map((room) => (
+            <RoomCard key={room.id} room={room} />
+          ))}
+        </ul>
       </section>
     </main>
   );
